[new] sr98::noralsy encoder, sr98::em4100 reader [fix] net::trust legacy flags [fix] dpapi decrypt by rpc, remove /system flag (incompatible with system key)
