+++ mimikatz/mimikatz.vcxproj.filters
@@ -281,6 +281,9 @@
     <ClCompile Include="..\modules\kull_m_ldap.c">
       <Filter>common modules</Filter>
     </ClCompile>
+    <ClCompile Include="modules\crypto\kuhl_m_crypto_pki.c">
+      <Filter>local modules\crypto</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="mimikatz.h" />
@@ -580,6 +583,9 @@
     </ClInclude>
     <ClInclude Include="..\modules\kull_m_ldap.h">
       <Filter>common modules</Filter>
+    </ClInclude>
+    <ClInclude Include="modules\crypto\kuhl_m_crypto_pki.h">
+      <Filter>local modules\crypto</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
