+++ mimikatz/modules/crypto/kuhl_m_crypto_sc.c
@@ -467,8 +467,6 @@ void kuhl_m_crypto_l_mdr(LPCWSTR szMdr, SCARDCONTEXT ctxScard, SCARDHANDLE hScar
 				enuma(&cd, NULL);
 				enuma(&cd, "mscp");
 				enuma(&cd, "mimikatz");
-				
-
 
 				status = cd.pfnCardDeleteContext(&cd);
 				if(status != SCARD_S_SUCCESS)
